add carousel for spotlight
