+++ locales/en.json
@@ -862,6 +862,7 @@
     "newestListHeading": "Newest List",
     "popularCollectionsHeading": "Popular Collections",
     "popularCollectionsDesc": "Discover the most popular collections on rmrk",
+    "spotlight": "Spotlight",
     "from": "from",
     "to": "to",
     "tx": {
